[Refactor 🎨](divine NotificationController): split NotificationController into interface and class
